added seed data and seed.js in seed folder. npm run seed to seed data after using db/schema.sql to initialise the database
