+++ .env.example
@@ -1,1 +1,4 @@
-PORT=''
+DB_NAME=''
+DB_USER=''
+DB_PASSWORD=''
+SESSION_SECRET=''
