+++ seed/chatData.json
@@ -1,0 +1,82 @@
+[
+  {
+    "user_a": 10,
+    "user_b": 9
+  },
+  {
+    "user_a": 3,
+    "user_b": 18
+  },
+  {
+    "user_a": 3,
+    "user_b": 10
+  },
+  {
+    "user_a": 17,
+    "user_b": 5
+  },
+  {
+    "user_a": 7,
+    "user_b": 2
+  },
+  {
+    "user_a": 1,
+    "user_b": 5
+  },
+  {
+    "user_a": 14,
+    "user_b": 3
+  },
+  {
+    "user_a": 12,
+    "user_b": 20
+  },
+  {
+    "user_a": 17,
+    "user_b": 8
+  },
+  {
+    "user_a": 9,
+    "user_b": 18
+  },
+  {
+    "user_a": 10,
+    "user_b": 7
+  },
+  {
+    "user_a": 5,
+    "user_b": 11
+  },
+  {
+    "user_a": 2,
+    "user_b": 16
+  },
+  {
+    "user_a": 7,
+    "user_b": 6
+  },
+  {
+    "user_a": 14,
+    "user_b": 17
+  },
+  {
+    "user_a": 8,
+    "user_b": 3
+  },
+  {
+    "user_a": 1,
+    "user_b": 13
+  },
+  {
+    "user_a": 7,
+    "user_b": 16
+  },
+  {
+    "user_a": 20,
+    "user_b": 15
+  },
+  {
+    "user_a": 20,
+    "user_b": 2
+  }
+]
