+++ seed/commentData.json
@@ -1,0 +1,142 @@
+[
+  {
+    "content": "Trip each they discussion other boy baby society stuff activity population significant beat they authority.",
+    "date_created": "2024-03-15T10:46:10.180886",
+    "likes": 47,
+    "user_id": 19,
+    "post_id": 12
+  },
+  {
+    "content": "Weight set else make sea nation top not kid action add.",
+    "date_created": "2024-02-07T10:46:10.180931",
+    "likes": 90,
+    "user_id": 13,
+    "post_id": 3
+  },
+  {
+    "content": "Special either floor project simple power image reason forget.",
+    "date_created": "2023-06-19T10:46:10.180958",
+    "likes": 72,
+    "user_id": 12,
+    "post_id": 16
+  },
+  {
+    "content": "Nearly available onto do sign hotel investment rich draw rate catch fly.",
+    "date_created": "2023-12-06T10:46:10.180984",
+    "likes": 80,
+    "user_id": 13,
+    "post_id": 5
+  },
+  {
+    "content": "Back reason explain alone will plan section decade probably provide.",
+    "date_created": "2023-09-06T10:46:10.181024",
+    "likes": 17,
+    "user_id": 15,
+    "post_id": 7
+  },
+  {
+    "content": "Clear key collection both yourself why concern peace prevent PM action with staff price fine.",
+    "date_created": "2024-02-24T10:46:10.181049",
+    "likes": 79,
+    "user_id": 13,
+    "post_id": 5
+  },
+  {
+    "content": "Husband far exactly base able similar message research more letter how.",
+    "date_created": "2024-01-05T10:46:10.181081",
+    "likes": 15,
+    "user_id": 16,
+    "post_id": 16
+  },
+  {
+    "content": "Husband whole there great head character force sometimes company opportunity.",
+    "date_created": "2023-10-06T10:46:10.181106",
+    "likes": 3,
+    "user_id": 3,
+    "post_id": 4
+  },
+  {
+    "content": "Late show range voice well tonight continue red style during new despite thought country become beautiful.",
+    "date_created": "2023-10-23T10:46:10.181136",
+    "likes": 49,
+    "user_id": 19,
+    "post_id": 10
+  },
+  {
+    "content": "North stop candidate form paper never sister travel officer participant program kid significant.",
+    "date_created": "2023-11-22T10:46:10.181170",
+    "likes": 84,
+    "user_id": 20,
+    "post_id": 8
+  },
+  {
+    "content": "Heart bed require drop want always more sit until there better several benefit product.",
+    "date_created": "2023-06-21T10:46:10.181198",
+    "likes": 47,
+    "user_id": 13,
+    "post_id": 7
+  },
+  {
+    "content": "Open position use home series decide large success.",
+    "date_created": "2024-03-06T10:46:10.181224",
+    "likes": 22,
+    "user_id": 6,
+    "post_id": 3
+  },
+  {
+    "content": "Appear catch country letter number design from return measure seek sort.",
+    "date_created": "2024-02-08T10:46:10.181251",
+    "likes": 68,
+    "user_id": 17,
+    "post_id": 12
+  },
+  {
+    "content": "Age audience audience performance father share him safe true authority federal a under later receive.",
+    "date_created": "2024-04-24T10:46:10.181290",
+    "likes": 22,
+    "user_id": 9,
+    "post_id": 13
+  },
+  {
+    "content": "Physical its condition clearly detail anything truth coach.",
+    "date_created": "2023-05-29T10:46:10.181315",
+    "likes": 31,
+    "user_id": 7,
+    "post_id": 3
+  },
+  {
+    "content": "Necessary pay wind production more full high only also that.",
+    "date_created": "2024-04-01T10:46:10.181341",
+    "likes": 3,
+    "user_id": 18,
+    "post_id": 11
+  },
+  {
+    "content": "Necessary check ask sell pay experience Republican fill management determine occur enter.",
+    "date_created": "2023-06-25T10:46:10.181366",
+    "likes": 30,
+    "user_id": 5,
+    "post_id": 13
+  },
+  {
+    "content": "Still nature red structure modern whether tonight.",
+    "date_created": "2024-03-06T10:46:10.181393",
+    "likes": 66,
+    "user_id": 17,
+    "post_id": 11
+  },
+  {
+    "content": "Whether wish government result method factor project American can natural candidate.",
+    "date_created": "2023-12-15T10:46:10.181419",
+    "likes": 49,
+    "user_id": 20,
+    "post_id": 8
+  },
+  {
+    "content": "Season maybe third for question environment available central contain citizen.",
+    "date_created": "2024-04-11T10:46:10.181445",
+    "likes": 56,
+    "user_id": 18,
+    "post_id": 19
+  }
+]
