+++ seed/mediaData.json
@@ -1,0 +1,162 @@
+[
+  {
+    "filename": "file_1.jpg",
+    "original_name": "original_file_1.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_1.jpg",
+    "size": 1514,
+    "post_id": 4
+  },
+  {
+    "filename": "file_2.jpg",
+    "original_name": "original_file_2.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_2.jpg",
+    "size": 8732,
+    "post_id": 8
+  },
+  {
+    "filename": "file_3.jpg",
+    "original_name": "original_file_3.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_3.jpg",
+    "size": 4580,
+    "post_id": 6
+  },
+  {
+    "filename": "file_4.jpg",
+    "original_name": "original_file_4.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_4.jpg",
+    "size": 6684,
+    "post_id": 3
+  },
+  {
+    "filename": "file_5.jpg",
+    "original_name": "original_file_5.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_5.jpg",
+    "size": 8089,
+    "post_id": 4
+  },
+  {
+    "filename": "file_6.jpg",
+    "original_name": "original_file_6.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_6.jpg",
+    "size": 3313,
+    "post_id": 6
+  },
+  {
+    "filename": "file_7.jpg",
+    "original_name": "original_file_7.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_7.jpg",
+    "size": 1608,
+    "post_id": 10
+  },
+  {
+    "filename": "file_8.jpg",
+    "original_name": "original_file_8.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_8.jpg",
+    "size": 4176,
+    "post_id": 19
+  },
+  {
+    "filename": "file_9.jpg",
+    "original_name": "original_file_9.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_9.jpg",
+    "size": 3832,
+    "post_id": 8
+  },
+  {
+    "filename": "file_10.jpg",
+    "original_name": "original_file_10.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_10.jpg",
+    "size": 9190,
+    "post_id": 14
+  },
+  {
+    "filename": "file_11.jpg",
+    "original_name": "original_file_11.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_11.jpg",
+    "size": 2182,
+    "post_id": 4
+  },
+  {
+    "filename": "file_12.jpg",
+    "original_name": "original_file_12.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_12.jpg",
+    "size": 2286,
+    "post_id": 9
+  },
+  {
+    "filename": "file_13.jpg",
+    "original_name": "original_file_13.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_13.jpg",
+    "size": 7239,
+    "post_id": 15
+  },
+  {
+    "filename": "file_14.jpg",
+    "original_name": "original_file_14.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_14.jpg",
+    "size": 4160,
+    "post_id": 2
+  },
+  {
+    "filename": "file_15.jpg",
+    "original_name": "original_file_15.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_15.jpg",
+    "size": 9987,
+    "post_id": 13
+  },
+  {
+    "filename": "file_16.jpg",
+    "original_name": "original_file_16.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_16.jpg",
+    "size": 3066,
+    "post_id": 6
+  },
+  {
+    "filename": "file_17.jpg",
+    "original_name": "original_file_17.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_17.jpg",
+    "size": 1564,
+    "post_id": 17
+  },
+  {
+    "filename": "file_18.jpg",
+    "original_name": "original_file_18.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_18.jpg",
+    "size": 9136,
+    "post_id": 9
+  },
+  {
+    "filename": "file_19.jpg",
+    "original_name": "original_file_19.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_19.jpg",
+    "size": 1567,
+    "post_id": 1
+  },
+  {
+    "filename": "file_20.jpg",
+    "original_name": "original_file_20.jpg",
+    "mimetype": "image/jpeg",
+    "path": "/uploads/file_20.jpg",
+    "size": 2691,
+    "post_id": 16
+  }
+]
