+++ seed/messageData.json
@@ -1,0 +1,142 @@
+[
+  {
+    "content": "This is a message 1.",
+    "date_created": "2024-05-27T10:27:03.734631",
+    "from_user": 11,
+    "to_user": 7,
+    "chat_id": 13
+  },
+  {
+    "content": "This is a message 2.",
+    "date_created": "2024-05-27T10:27:03.734648",
+    "from_user": 16,
+    "to_user": 14,
+    "chat_id": 16
+  },
+  {
+    "content": "This is a message 3.",
+    "date_created": "2024-05-27T10:27:03.734652",
+    "from_user": 11,
+    "to_user": 3,
+    "chat_id": 18
+  },
+  {
+    "content": "This is a message 4.",
+    "date_created": "2024-05-27T10:27:03.734656",
+    "from_user": 8,
+    "to_user": 14,
+    "chat_id": 12
+  },
+  {
+    "content": "This is a message 5.",
+    "date_created": "2024-05-27T10:27:03.734660",
+    "from_user": 15,
+    "to_user": 8,
+    "chat_id": 12
+  },
+  {
+    "content": "This is a message 6.",
+    "date_created": "2024-05-27T10:27:03.734663",
+    "from_user": 3,
+    "to_user": 12,
+    "chat_id": 15
+  },
+  {
+    "content": "This is a message 7.",
+    "date_created": "2024-05-27T10:27:03.734667",
+    "from_user": 8,
+    "to_user": 19,
+    "chat_id": 8
+  },
+  {
+    "content": "This is a message 8.",
+    "date_created": "2024-05-27T10:27:03.734671",
+    "from_user": 1,
+    "to_user": 19,
+    "chat_id": 14
+  },
+  {
+    "content": "This is a message 9.",
+    "date_created": "2024-05-27T10:27:03.734675",
+    "from_user": 1,
+    "to_user": 17,
+    "chat_id": 5
+  },
+  {
+    "content": "This is a message 10.",
+    "date_created": "2024-05-27T10:27:03.734679",
+    "from_user": 20,
+    "to_user": 3,
+    "chat_id": 7
+  },
+  {
+    "content": "This is a message 11.",
+    "date_created": "2024-05-27T10:27:03.734683",
+    "from_user": 11,
+    "to_user": 7,
+    "chat_id": 11
+  },
+  {
+    "content": "This is a message 12.",
+    "date_created": "2024-05-27T10:27:03.734686",
+    "from_user": 12,
+    "to_user": 10,
+    "chat_id": 14
+  },
+  {
+    "content": "This is a message 13.",
+    "date_created": "2024-05-27T10:27:03.734691",
+    "from_user": 5,
+    "to_user": 3,
+    "chat_id": 11
+  },
+  {
+    "content": "This is a message 14.",
+    "date_created": "2024-05-27T10:27:03.734696",
+    "from_user": 1,
+    "to_user": 6,
+    "chat_id": 12
+  },
+  {
+    "content": "This is a message 15.",
+    "date_created": "2024-05-27T10:27:03.734699",
+    "from_user": 17,
+    "to_user": 17,
+    "chat_id": 6
+  },
+  {
+    "content": "This is a message 16.",
+    "date_created": "2024-05-27T10:27:03.734703",
+    "from_user": 15,
+    "to_user": 10,
+    "chat_id": 19
+  },
+  {
+    "content": "This is a message 17.",
+    "date_created": "2024-05-27T10:27:03.734707",
+    "from_user": 7,
+    "to_user": 9,
+    "chat_id": 9
+  },
+  {
+    "content": "This is a message 18.",
+    "date_created": "2024-05-27T10:27:03.734710",
+    "from_user": 6,
+    "to_user": 4,
+    "chat_id": 8
+  },
+  {
+    "content": "This is a message 19.",
+    "date_created": "2024-05-27T10:27:03.734714",
+    "from_user": 12,
+    "to_user": 20,
+    "chat_id": 18
+  },
+  {
+    "content": "This is a message 20.",
+    "date_created": "2024-05-27T10:27:03.734718",
+    "from_user": 7,
+    "to_user": 7,
+    "chat_id": 16
+  }
+]
