+++ seed/userData.json
@@ -1,0 +1,55 @@
+[
+  {
+    "username": "shawnlittle",
+    "email": "johnsonpamela@fox.biz",
+    "password": "$2b$12$UZtzPnCTUz82gtn.21kKzOj8JH95uWY26AxsvcMX5qUr53WYgHMbu"
+  },
+  {
+    "username": "owagner",
+    "email": "stephanie69@duffy-baldwin.com",
+    "password": "$2b$12$mCG86huTIU4eRcwFOSCp2OtFs5UH96/eR.3wLyIqBEFGc3PnFIVLa"
+  },
+  {
+    "username": "nathan63",
+    "email": "lauren49@adams-parks.org",
+    "password": "$2b$12$5rkf2UiLl62PB1rIZK1BaOEwflcyrLb.Um3F2Ro1LFNVzDqz9GfK."
+  },
+  {
+    "username": "caroline43",
+    "email": "sandy01@hotmail.com",
+    "password": "$2b$12$A46e2/KCI8z1hbi90heeD.yF9UqKZPPDYpB.T1zyqasFuElkyQaKe"
+  },
+  {
+    "username": "jaredcastillo",
+    "email": "ythomas@gmail.com",
+    "password": "$2b$12$gXzoj1veTkyD2CVNBidWXeJEH2Q7ajvBXJ1K6apc1GjKtResdmws2"
+  },
+  {
+    "username": "harrisrhonda",
+    "email": "katelynlewis@gmail.com",
+    "password": "$2b$12$IG8qn83n05WkvWLgyTuErusLdIcRTdfcOLmAgWfapy1CCfBQy3Qiq"
+  },
+  {
+    "username": "kaitlin58",
+    "email": "dyergeorge@gmail.com",
+    "password": "$2b$12$Y5.wySDm1K6Hs2leQEzvqeYkbRw7a8Wl0e9QW4p/qQ.3ryaHflYWm"
+  },
+  {
+    "username": "christinescott",
+    "email": "brandon88@gmail.com",
+    "password": "$2b$12$YaMyV8iA1VAge0SvC6tDSeWSb6ubRRnqwgUBOhumsc2dx9AKcmYPO"
+  },
+  {
+    "username": "brownbruce",
+    "email": "trodriguez@stein.com",
+    "password": "$2b$12$GJJS2ARAi/EO1lDqrBKi.eQ4TTRbSj5TgOIShrfvBPCETcCU17pNG"
+  },
+  {
+    "username": "keith41",
+    "email": "thomas74@gmail.com",
+    "password": "$2b$12$aHJUTaXbSj8JNl33PiruauBIglb3jhuHDkdelldiLmy/LI81bh37m"
+  },
+  {
+    "username": "kdavis",
+    "email": "lozanodawn@marshall.info",
+    "password": "$2b$12$rDrPyhvPoB.0MPQ0AhGDOOTbuHL5iiUdFl3Wawg8DPi3eUCZoX
